# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kiali project, observability for the Istio service mesh",
    "title": "Kiali",
    "version": "_"
  },
  "basePath": "/api",
  "paths": {
...
    "/auth/info": {
      "get": {
        "description": "Endpoint to get login info, such as strategy, authorization endpoints\nfor OAuth providers and so on.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "AuthenticationInfo",
        "responses": {
          "200": {
            "$ref": "#/responses/authenticationInfo"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
    "/authenticate": {
      "get": {
        "security": [
          {
            "authorization": [
              "user",
              "password"
            ]
          }
        ],
        "description": "Endpoint to authenticate the user",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "Authenticate",
        "responses": {
          "200": {
            "$ref": "#/responses/tokenResponse"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      },
      "post": {
        "description": "Endpoint to check if a token from Openshift is working correctly",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "OpenshiftCheckToken",
        "responses": {
          "200": {
            "$ref": "#/responses/tokenResponse"
          },
          "500": {
            "$ref": "#/responses/internalError"
          }
        }
      }
    },
...
    "/grafana": {
      "get": {
        "description": "Get the grafana URL and other descriptors",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "operationId": "grafanaInfo",
        "responses": {
          "200": {
            "$ref": "#/responses/grafanaInfoResponse"
          },
          "204": {
            "$ref": "#/responses/noContent"
          },
          "500": {
            "$ref": "#/responses/internalError"
          },
          "503": {
            "$ref": "#/responses/serviceUnavailableError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "GrafanaInfo": {
      "description": "GrafanaInfo provides information to access Grafana dashboards",
      "type": "object",
      "properties": {
        "serviceDashboardPath": {
          "type": "string",
          "x-go-name": "ServiceDashboardPath"
        },
        "url": {
          "type": "string",
          "x-go-name": "URL"
        },
        "varNamespace": {
          "type": "string",
          "x-go-name": "VarNamespace"
        },
        "varService": {
          "type": "string",
          "x-go-name": "VarService"
        },
        "varWorkload": {
          "type": "string",
          "x-go-name": "VarWorkload"
        },
        "workloadDashboardPath": {
          "type": "string",
          "x-go-name": "WorkloadDashboardPath"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/models"
    },
...
    "TokenResponse": {
      "description": "This is used for returning the token",
      "type": "object",
      "title": "TokenResponse tokenResponse",
      "required": [
        "username",
        "token",
        "expiresOn"
      ],
      "properties": {
        "expiresOn": {
          "description": "The expired time for the token\nA string with the Datetime when the token will be expired",
          "type": "string",
          "x-go-name": "ExpiresOn",
          "example": "Thu, 07 Mar 2019 17:50:26 +0000"
        },
        "token": {
          "description": "The authentication token\nA string with the authentication token for the user",
          "type": "string",
          "x-go-name": "Token",
          "example": "zI1NiIsIsR5cCI6IkpXVCJ9.ezJ1c2VybmFtZSI6ImFkbWluIiwiZXhwIjoxNTI5NTIzNjU0fQ.PPZvRGnR6VA4v7FmgSfQcGQr-VD"
        },
        "username": {
          "description": "The username for the token\nA string with the user's username",
          "type": "string",
          "x-go-name": "Username",
          "example": "admin"
        }
      },
      "x-go-package": "github.com/mayadata-io/kiali/handlers"
    },
...
  },
  "responses": {
...
    "authenticationInfo": {
      "description": "Return the information necessary to handle login",
      "headers": {
        "AuthorizationEndpoint": {
          "type": "string"
        },
        "Strategy": {
          "type": "string"
        }
      }
    },
...
    "grafanaInfoResponse": {
      "description": "Return all the descriptor data related to Grafana",
      "schema": {
        "$ref": "#/definitions/GrafanaInfo"
      }
    },
...
    "internalError": {
      "description": "A Internal is the error message that means something has gone wrong",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 500,
            "x-go-name": "Code",
            "example": 500
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "noContent": {
      "description": "NoContent: the response is empty",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 204,
            "x-go-name": "Code",
            "example": 204
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "serviceUnavailableError": {
      "description": "A Internal is the error message that means something has gone wrong",
      "schema": {
        "type": "object",
        "properties": {
          "code": {
            "description": "HTTP status code",
            "type": "integer",
            "format": "int32",
            "default": 503,
            "x-go-name": "Code",
            "example": 503
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
    "tokenResponse": {
      "description": "HTTP status code 200 and tokenGenerated model in data",
      "schema": {
        "$ref": "#/definitions/TokenResponse"
      }
    },
...
  }
}